+++ adapter-unityads/CloudXUnityAdsAdapter.xcframework/ios-arm64/CloudXUnityAdsAdapter.framework/Headers/CLXUnityAdsInterstitialFactory.h
@@ -1,0 +1,22 @@
+//
+//  CLXUnityAdsInterstitialFactory.h
+//  CloudXUnityAdsAdapter
+//
+
+#import <Foundation/Foundation.h>
+
+#if __has_include(<CloudXCore/CloudXCore.h>)
+#import <CloudXCore/CloudXCore.h>
+#else
+@import CloudXCore;
+#endif
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface CLXUnityAdsInterstitialFactory : NSObject <CLXAdapterInterstitialFactory>
+
++ (instancetype)createInstance;
+
+@end
+
+NS_ASSUME_NONNULL_END