+++ core/CloudXCore.xcframework/ios-arm64_x86_64-simulator/CloudXCore.framework/Headers/CLXAdNetwork.h
@@ -17,7 +17,7 @@ extern CLXAdNetwork const CLXAdNetworkCloudX;
 extern CLXAdNetwork const CLXAdNetworkCloudXRenderer;
 extern CLXAdNetwork const CLXAdNetworkVungle;
 extern CLXAdNetwork const CLXAdNetworkInMobi;
-extern CLXAdNetwork const CLXAdNetworkUnity;
+extern CLXAdNetwork const CLXAdNetworkUnityAds;
 extern CLXAdNetwork const CLXAdNetworkMoloco;
 
 /// All known ad networks.